Convert wiki-style links

+++ archive/lexicon1/build.bash
@@ -1,8 +1,10 @@
 for x in *.txt; do
-  mv "$x" "${x%.txt}.md"
+  [ -f "$x" ] && mv "$x" "${x%.txt}.md"
 done
-mv LexiconOne.md index.md
+[ -f LexiconOne.md ] && mv -i LexiconOne.md index.md
 
-# Replace [[The Dog]] with [The Dog](The Dog)
-
-# Replace [The Loping Doge](https://lexicon.za3k.com/index.php/Loping%20Doge,%20The) with [The Loping Doge](Loping Doge, The)
+for x in *.md; do
+  # Replace [[The Dog]] with [The Dog](The Dog)
+  sed --in-place -Ee "s/\[\[(.*)\]\]/[\1](\1)/" "$x"
+  # Replace [The Loping Doge](https://lexicon.za3k.com/index.php/Loping%20Doge,%20The) with [The Loping Doge](Loping Doge, The)
+done
